+++ ChangeLog.md
@@ -1,4 +1,7 @@
 # Prophets of the Labyrinth Change Log
+#### Prophets of the Labyrinth Version 0.9.0:
+- New slash command: `/armory` that allows checking stats on a weapon given by name
+- The changelog no longer offers "premimum features" for sponsoring on GitHub
 #### Prophets of the Labyrinth Version 0.8.0:
 - Added "Data Policy" page to manual
 - Enemy hp now scales with party size